#500 Cleanup old docs setup
