+++ docs/layout/partials/header.html
@@ -81,7 +81,7 @@ click(<span class="hljs-string">"Google Search"</span>)</code></pre>
     </p>
     <p>You can see the browser performing these actions as you type and press enter for each command.
     </p>
-    <img src="./asset/images/usage.gif" class="img-responsive" />
+    <img src="./assets/images/usage.gif" class="img-responsive" />
     <p>Taiko’s REPL keeps a history of all successful commands. Once you finish a flow of execution, you can generate a test script using the special command <span class="highlight">.code</span>
     </p>
