--- docs/layoutOld/helpers/lowerCase.js
@@ -1,4 +1,0 @@
-var Handlebars = require('handlebars');
-module.exports = function(text){
-    return text.toLowerCase();
-};
