--- docs/layoutOld/page.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, minimal-ui">
-    <title>{{title}}</title>
-    <link type="text/css" rel="stylesheet" href=".asset/css/github-markdown.css">
-    <link type="text/css" rel="stylesheet" href=".asset/css/pilcrow.css">
-    <link type="text/css" rel="stylesheet" href=".asset/css/hljs-github.min.css"/>
-    <link type="text/css" rel="stylesheet" href=".asset/css/style.css"/>
-  </head>
-  <body>
-    <article class="markdown-body">
-      {{~> toc}}
-      {{~> content}}
-    </article>
-  </body>
-</html>
