--- docs/layoutOld/partials/toc.html
@@ -1,14 +1,0 @@
-<h1><a href="http://taiko.gauge.org">Taiko</a></h1>
-<h3>Table of Contents</h3>
-<ul class="nav nav-list">
-    {{#each toc}}
-        <div style="margin-bottom: 2%"> 
-        <li style="font-weight: bold; font-size: 18px;">{{@key}}</li>
-            <span style="font-family: monospace;">
-            {{#each this}}
-                <a href=#{{{lowerCase this}}}>{{this}}</a>{{#unless @last}},{{/unless}}
-            {{/each}}
-            </span>
-        </div>     
-    {{/each}}
-</ul>
