+++ docs/setup.js
@@ -1,30 +1,34 @@
-const fs = require('fs');
-const mds = require('markdown-styles');
+const marked = require('marked');
+const path = require('path');
+const handlebars = require('handlebars');
+const fs = require('fs-extra');
 const { metadata } = require('../lib/taiko');
 
-const content = `
-title: Taiko
-toc: ${JSON.stringify(metadata)}
----
-` + fs.readFileSync('docs/index.md', 'utf-8');
-fs.writeFileSync('docs/index.md', content, 'utf-8');
-
-function generateDoc(layout) {
-    mds.render(mds.resolveArgs({
-        input: 'docs/index.md',
-        output: 'docs/site',
-        layout: layout,
-    }), function () {
-        console.log('All done!');
-    });
-}
-
-function main(arg) {
-    if (arg === '--new') {
-        generateDoc('docs/layout');
-    }else{
-        generateDoc('docs/layoutOld');
+var createDir = (dir) => {
+    if (!fs.existsSync(dir)){
+        fs.mkdirSync(dir);
     }
-}
+};
 
-main(process.argv[2]);
+createDir('docs/site/');
+createDir('docs/site/assets');
+console.log('Copying assets...');
+fs.copySync('docs/layout/assets','docs/site/assets');
+
+console.log('Converting markdown to partial...');
+var renderer = new marked.Renderer();
+fs.writeFileSync('docs/layout/partials/content.html', marked(fs.readFileSync('docs/index.md', 'utf8'), { renderer: renderer }), 'utf-8');
+
+console.log('Compiling partials to site...');
+fs.readdirSync('docs/layout/partials').forEach(file => {
+    var partialName = path.basename(file, path.extname(file));
+    handlebars.registerPartial(partialName, fs.readFileSync('docs/layout/partials/'+file, 'utf8'));
+});
+handlebars.registerHelper('lowerCase', require('./layout/helpers/lowerCase'));
+
+var template = handlebars.compile(fs.readFileSync('docs/layout/page.html', 'utf8'));
+var html    = template({toc:metadata});
+
+
+console.log('Writing to index.html...');
+fs.writeFileSync('docs/site/index.html', html, 'utf8');
