+++ package.json
@@ -8,7 +8,6 @@
 	},
 	"scripts": {
 		"lint": "eslint bin/*.js lib/*.js test/*/*.js",
-		"doc-new": "documentation build lib/taiko.js -f md --shallow -o docs/index.md --markdown-toc=false && node docs/setup.js --new",
 		"doc": "documentation build lib/taiko.js -f md --shallow -o docs/index.md --markdown-toc=false && node docs/setup.js",
 		"test:api": "node test/unit-tests/taiko-test.js",
 		"examples": "npm install --prefix examples && npm test --prefix examples",
